+++ core/Cargo.toml
@@ -42,6 +42,7 @@ regress = "0.4"
 flash-lso = { git = "https://github.com/ruffle-rs/rust-flash-lso", rev = "19fecd07b9888c4bdaa66771c468095783b52bed" }
 json = "0.12.4"
 lzma-rs = {version = "0.2.0", optional = true }
+dasp = { git = "https://github.com/RustAudio/dasp", rev = "f05a703", features = ["interpolate", "interpolate-linear", "signal"] }
 
 [dependencies.jpeg-decoder]
 version = "0.1.22"
