+++ desktop/Cargo.toml
@@ -14,7 +14,6 @@ ruffle_render_wgpu = { path = "../render/wgpu", features = ["clap"] }
 env_logger = "0.9.0"
 generational-arena = "0.2.8"
 log = "0.4"
-dasp = { git = "https://github.com/RustAudio/dasp", rev = "f05a703", features = ["interpolate", "interpolate-linear", "signal"] }
 winit = "0.25.0"
 webbrowser = "0.5.5"
 url = "2.2.2"
